+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [0.2.1] - 2026-02-09
+
+### Added
+
+- ContextService to use short-term storage for add/get/clear 
+
+### Fixed
+
+- Ensure short-term memory starts with a user message to avoid broken tool call/result pairs.
+
+## [0.2.0] - 2026-02-06
+
+### Added
+
+- Support for custom json schema when using LLM Extractor
+
 ## [0.1.0] - 2026-02-03
 
 Initial Release